factor out signing
