+++ src/services/collectionService.ts
@@ -1,18 +1,7 @@
 import { Request, Response } from 'express';
-import httpSignature from 'http-signature';
 import { getOutboxFromDB, addNoteToDB } from '../dbService';
 import { Note } from '../types';
-
-const signActivity = (activity: any, privateKey: string, keyId: string): any => {
-  const signedActivity = { ...activity };
-  const options = {
-    key: privateKey,
-    keyId: keyId,
-    headers: ['(request-target)', 'date', 'digest'],
-  };
-  httpSignature.sign(signedActivity, options);
-  return signedActivity;
-};
+import { signActivity } from './utils';
 
 export const getOutbox = async (req: Request, res: Response): Promise<void> => {
   const username = req.params.username;
